// src/app/core/models/decision.ts
import { BaseModel } from './baseModel';
import { Attachment } from './attachment';

export class Decision extends BaseModel {
    id: number;
    meeting_id: number;
    agenda_id: number;
    vote_type_id: number;
    is_started: boolean;
    vote_subject_ar: string;
    vote_subject_en: string;
    vote_schedule_from: any;
    vote_schedule_to: any;
    agenda_title_ar:  string;
    agenda_title_en: string;
    meeting_title_ar: string;
    meeting_title_en: string;
    vote_type_name_ar: string;
    vote_type_name_en: string;
    decision_type_name_ar: string;
    decision_type_name_en: string;
    decision_due_date: any;
    vote_users_ids: Array<any>;
    committee_id: number;
    is_secret: boolean = false;
    attachments: Array<Attachment> = [];
    creation_date: any;
    creator_name_ar: string;
    creator_name: string;
    vote_result_status_name_en: string;
    vote_result_status_name_ar: string;
    vote_description: string;
    vote_result_status_id: number;
    votes_number: number;
    voters: Array<any> = [];
    can_edit: boolean;
    can_vote: boolean;
    yes_votes: number;
    no_votes: number;
    decision_type_id: number;
    abstained_votes: number;
    organization_image_url:  string;
    creator_image_url:  string;
    creation_same_day : boolean;
    show_due_date : boolean;
    creation_hour_diff : number;
    creation_minute_diff : number;
    due_months: number;
    due_days: number;
    due_hours: number;
    due_minutes : number;
    tasks: Array<any> = [];
	is_signed: boolean;
	is_creator: boolean;
}